20-100x Speed Optimization: tf.browser.toPixels()
This fix enables creation of pixel data to be faster by using lower
level bit shifting with a Uint32Array.  Leads to a performance boost of
roughly 20 - 100x depending on device used vs old method.  Assumes
little endian, so check is also added for this. If not little endian we
fall back to the original method. Supports depths 1, 3, and 4 image
tensors whose values are normalized between 0 - 1, or regular 0 - 255
values.